+++ adapters/DuAd/DuAdAdapter/GADDuAdError.m
@@ -13,10 +13,11 @@
 // limitations under the License.
 
 #import "GADDuAdError.h"
+#import "GADMAdapterDuAdConstants.h"
 
 NSError *GADDUErrorWithDescription(NSString *description) {
   NSDictionary *userInfo =
       @{NSLocalizedDescriptionKey : description, NSLocalizedFailureReasonErrorKey : description};
-  NSError *error = [NSError errorWithDomain:@"com.google.mediation.duad" code:0 userInfo:userInfo];
+  NSError *error = [NSError errorWithDomain:kGADMAdapterDuAdErrorDomain code:0 userInfo:userInfo];
   return error;
 }