+++ adapters/Facebook/FacebookAdapter/GADFBAdapterDelegate.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-@import Foundation;
-@import FBAudienceNetwork;
+#import <Foundation/Foundation.h>
+#import <FBAudienceNetwork/FBAudienceNetwork.h>
 
 @protocol GADMAdNetworkAdapter;
 @protocol GADMAdNetworkConnector;