+++ adapters/Facebook/FacebookAdapter/GADFBError.h
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-@import Foundation;
+#import <Foundation/Foundation.h>
 
 /// Macro wrapper for NSLog only if debug mode is enabled.
 #if DEBUG