+++ adapters/Facebook/FacebookAdapter/OpenBidding/GADFBBannerRenderer.m
@@ -16,8 +16,8 @@
 #import "GADFBError.h"
 #import "GADMAdapterFacebookConstants.h"
 #import "GADMediationAdapterFacebook.h"
-@import FBAudienceNetwork;
-@import AdSupport;
+#import <FBAudienceNetwork/FBAudienceNetwork.h>
+#import <AdSupport/AdSupport.h>
 
 @interface GADFBBannerRenderer () <GADMediationBannerAd, FBAdViewDelegate> {
   // The completion handler to call when the ad loading succeeds or fails.