+++ adapters/Unity/UnityAdapter/GADMAdapterUnitySingleton.h
@@ -25,7 +25,8 @@
 
 /// Configures a reward-based video ad with provided |gameID| and |adapterDelegate| and returns
 /// YES if successful; otherwise returns NO.
-- (void)configureRewardedAdWithGameID:(NSString *)gameID;
+
+- (void)initializeWithGameID:(NSString *)gameID;
 
 /// Requests a reward-based video ad with |adapterDelegate|.
 - (void)requestRewardedAdWithDelegate:
@@ -38,10 +39,8 @@
                                           adapterDelegate;
 
 /// Configures an interstitial ad with provided |gameID| and |adapterDelegate|.
-- (void)configureInterstitialAdWithGameID:(NSString *)gameID
-                                 delegate:
-                                     (id<GADMAdapterUnityDataProvider, UnityAdsExtendedDelegate>)
-                                         adapterDelegate;
+- (void)requestInterstitialAdWithDelegate:
+    (id<GADMAdapterUnityDataProvider, UnityAdsExtendedDelegate>)adapterDelegate;
 
 /// Presents an interstitial ad for |viewController| with |adapterDelegate|.
 - (void)presentInterstitialAdForViewController:(UIViewController *)viewController